+++ src/BusinessLogic/Country/CountryService.php
@@ -123,18 +123,6 @@ class CountryService extends BaseService
     );
 
     /**
-     * Returns whether the country with provided ISO code is in a list of supported countries.
-     *
-     * @param string $isoCode Two-letter country code.
-     *
-     * @return bool
-     */
-    public function isCountrySupported($isoCode)
-    {
-        return array_key_exists($isoCode, static::$supportedCountries);
-    }
-
-    /**
      * Checks if given country is one of the four base countries ('ES', 'DE', 'FR', 'IT').
      *
      * @param string $countryCode Country ISO-2 code.
@@ -149,7 +137,7 @@ class CountryService extends BaseService
     /**
      * Returns a list of supported country DTOs.
      *
-     * @return \Packlink\BusinessLogic\Country\Country[]
+     * @return RegistrationCountry[]
      *
      * @noinspection PhpUnhandledExceptionInspection
      * @noinspection PhpDocMissingThrowsInspection
@@ -167,7 +155,7 @@ class CountryService extends BaseService
                     $country['registration_link']
                 ) . '&platform=PRO';
 
-            $countries[$country['code']] = FrontDtoFactory::get(Country::CLASS_KEY, $country);
+            $countries[$country['code']] = FrontDtoFactory::get(RegistrationCountry::CLASS_KEY, $country);
         }
 
         return $countries;