+++ src/BusinessLogic/ShippingMethod/ShippingCostCalculator.php
@@ -2,12 +2,16 @@
 
 namespace Packlink\BusinessLogic\ShippingMethod;
 
+use Exception;
+use InvalidArgumentException;
 use Logeecom\Infrastructure\Http\Exceptions\HttpBaseException;
+use Logeecom\Infrastructure\Logger\Logger;
 use Logeecom\Infrastructure\ServiceRegister;
 use Packlink\BusinessLogic\Http\DTO\Package;
 use Packlink\BusinessLogic\Http\DTO\ShippingServiceDetails;
 use Packlink\BusinessLogic\Http\DTO\ShippingServiceSearch;
 use Packlink\BusinessLogic\Http\Proxy;
+use Packlink\BusinessLogic\PostalCode\PostalCodeTransformer;
 use Packlink\BusinessLogic\ShippingMethod\Exceptions\FixedPriceValueOutOfBoundsException;
 use Packlink\BusinessLogic\ShippingMethod\Models\ShippingMethod;
 use Packlink\BusinessLogic\ShippingMethod\Models\ShippingService;
@@ -83,9 +87,8 @@ class ShippingCostCalculator
 
         $result = array();
         $package = self::preparePackages($packages);
-        $params = new ShippingServiceSearch(null, $fromCountry, $fromZip, $toCountry, $toZip, array($package));
         try {
-            $response = self::getProxy()->getShippingServicesDeliveryDetails($params);
+            $response = self::getPacklinkServices($fromCountry, $fromZip, $toCountry, $toZip, $package);
 
             $result = self::calculateShippingCostsPerShippingMethod(
                 $shippingMethods,
@@ -112,7 +115,7 @@ class ShippingCostCalculator
     /**
      * Returns cheapest service in shipping method.
      *
-     * @param \Packlink\BusinessLogic\ShippingMethod\Models\ShippingMethod $method
+     * @param ShippingMethod $method
      * @param string $fromCountry From country code.
      * @param string $fromZip From zip code.
      * @param string $toCountry To country code.
@@ -130,13 +133,10 @@ class ShippingCostCalculator
         array $packages
     ) {
         $package = self::preparePackages($packages);
-        $searchParams = new ShippingServiceSearch(null, $fromCountry, $fromZip, $toCountry, $toZip, array($package));
 
-        /** @var Proxy $proxy */
-        $proxy = ServiceRegister::getService(Proxy::CLASS_NAME);
         try {
-            $services = $proxy->getShippingServicesDeliveryDetails($searchParams);
-        } catch (\Exception $e) {
+            $services = self::getPacklinkServices($fromCountry, $fromZip, $toCountry, $toZip, $package);
+        } catch (Exception $e) {
             $services = array();
         }
 
@@ -167,7 +167,7 @@ class ShippingCostCalculator
             return $result;
         }
 
-        throw new \InvalidArgumentException(
+        throw new InvalidArgumentException(
             'No service is available for shipping method '
             . $method->getId() . ' for given destination country ' . $toCountry
             . ' and given packages ' . json_encode($packages)
@@ -175,9 +175,68 @@ class ShippingCostCalculator
     }
 
     /**
+     * Returns available Packlink services for the given departure and destination information.
+     * Transforms destination postal code to a supported format, if the country is supported for postal code transformation.
+     * In case of a 400 response, the request to Packlink API is reattempted with original destination postal code.
+     *
+     * @param string $fromCountry Departure country code.
+     * @param string $fromZip Departure zip code.
+     * @param string $toCountry Destination country code.
+     * @param string $toZip Destination zip code.
+     * @param Package $package
+     *
+     * @return array
+     *
+     * @throws HttpBaseException
+     * @throws \Logeecom\Infrastructure\Http\Exceptions\HttpAuthenticationException
+     * @throws \Logeecom\Infrastructure\Http\Exceptions\HttpCommunicationException
+     * @throws \Logeecom\Infrastructure\Http\Exceptions\HttpRequestException
+     */
+    protected static function getPacklinkServices($fromCountry, $fromZip, $toCountry, $toZip, $package)
+    {
+        $transformedPostalCode = '';
+
+        try {
+            $transformedPostalCode = PostalCodeTransformer::transform($toCountry, $toZip);
+            $searchParams = new ShippingServiceSearch(
+                null,
+                $fromCountry,
+                $fromZip,
+                $toCountry,
+                $transformedPostalCode,
+                array($package)
+            );
+
+            return self::getProxy()->getShippingServicesDeliveryDetails($searchParams);
+        } catch (HttpBaseException $e) {
+            if ($toZip !== $transformedPostalCode
+                && $e->getCode() === 400
+                && $e->getMessage() === 'Location not valid for the input data'
+            ) {
+                Logger::logWarning("Request with transformed postal code $transformedPostalCode for $toCountry country resulted in a unsuccessful response.");
+
+                $searchParams = new ShippingServiceSearch(
+                    null,
+                    $fromCountry,
+                    $fromZip,
+                    $toCountry,
+                    $toZip,
+                    array($package)
+                );
+
+                return self::getProxy()->getShippingServicesDeliveryDetails($searchParams);
+            }
+
+            throw $e;
+        } catch (InvalidArgumentException $e) {
+            return array();
+        }
+    }
+
+    /**
      * Calculates shipping method cost based on given criteria.
      *
-     * @param \Packlink\BusinessLogic\ShippingMethod\Models\ShippingMethod $shippingMethod
+     * @param ShippingMethod $shippingMethod
      * @param float $totalAmount
      *
      * @param int $serviceId
@@ -393,7 +452,7 @@ class ShippingCostCalculator
     }
 
     /**
-     * @param \Packlink\BusinessLogic\ShippingMethod\Models\ShippingMethod $method
+     * @param ShippingMethod $method
      * @param $totalAmount
      * @param $totalWeight
      *