+++ tests/BusinessLogic/Country/WarehouseCountryServiceTest.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace BusinessLogic\Country;
+
+use Logeecom\Infrastructure\ServiceRegister;
+use Logeecom\Tests\BusinessLogic\Common\BaseTestWithServices;
+use Packlink\BusinessLogic\Country\WarehouseCountryService;
+
+/**
+ * Class WarehouseCountryServiceTest
+ *
+ * @package BusinessLogic\Country
+ */
+class WarehouseCountryServiceTest extends BaseTestWithServices
+{
+    public function testGetSupportedCountries()
+    {
+        /** @var WarehouseCountryService $service */
+        $service = ServiceRegister::getService(WarehouseCountryService::CLASS_NAME);
+
+        $countries = $service->getSupportedCountries();
+
+        $this->assertNotEmpty($countries);
+        $this->assertCount(22, $countries);
+        $this->assertArrayHasKey('MX', $countries);
+        $this->assertEquals('Mexico', $countries['MX']->name);
+        $this->assertEquals('MX', $countries['MX']->code);
+        $this->assertEquals('21900', $countries['MX']->postalCode);
+    }
+
+    public function testSupportedCountry()
+    {
+        /** @var WarehouseCountryService $service */
+        $service = ServiceRegister::getService(WarehouseCountryService::CLASS_NAME);
+
+        $this->assertTrue($service->isCountrySupported('CZ'));
+    }
+
+    public function testUnsupportedCountry()
+    {
+        /** @var WarehouseCountryService $service */
+        $service = ServiceRegister::getService(WarehouseCountryService::CLASS_NAME);
+
+        $this->assertFalse($service->isCountrySupported('RS'));
+    }
+}